requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.13.0; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.11.0; python_version >= '3.9'
attrs==25.4.0; python_version >= '3.9'
beautifulsoup4==4.14.2; python_full_version >= '3.7.0'
cachetools==6.2.0; python_version >= '3.9'
certifi==2025.10.5; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
deep-translator==1.11.4; python_version >= '3.7' and python_version < '4.0'
discord.py[voice]==2.6.4; python_version >= '3.8'
feedparser==6.0.12; python_version >= '3.6'
ffmpeg==1.4
frozenlist==1.8.0; python_version >= '3.9'
google-ai-generativelanguage==0.6.15; python_version >= '3.7'
google-api-core[grpc]==2.25.2; python_version >= '3.7'
google-api-python-client==2.184.0; python_version >= '3.7'
google-auth==2.41.1; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-generativeai==0.8.5; python_version >= '3.9'
googleapis-common-protos==1.70.0; python_version >= '3.7'
googletrans==4.0.2; python_version >= '3.8'
grpcio==1.75.1; python_version >= '3.9'
grpcio-status==1.71.2; python_version >= '3.9'
gtts==2.5.4; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
h2==4.3.0; python_version >= '3.9'
hpack==4.1.0; python_version >= '3.9'
httpcore==1.0.9; python_version >= '3.8'
httplib2==0.31.0; python_version >= '3.6'
httpx[http2]==0.28.1; python_version >= '3.8'
hyperframe==6.1.0; python_version >= '3.9'
idna==3.11; python_version >= '3.8'
multidict==6.7.0; python_version >= '3.9'
numpy==2.3.3; python_version >= '3.11'
propcache==0.4.1; python_version >= '3.9'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.5; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pyaudio==0.2.14
pycparser==2.23; python_version >= '3.8'
pydantic==2.11.10; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydub==0.25.1
pynacl==1.5.0; python_version >= '3.6'
pyparsing==3.2.5; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
sgmllib3k==1.0.0
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.8; python_version >= '3.9'
speechrecognition==3.14.3; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
wikipedia-api==0.8.1
yarl==1.22.0; python_version >= '3.9'
yt-dlp==2025.9.26; python_version >= '3.9'
